platform/translator/pyproject.toml
[project]
name = "sohn-translator"
version = "0.1.0"
description = "Unblinded-Formula translator: PDF -> chunks -> RAG -> Sohn LLM -> 7-col CSV"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pdfplumber>=0.11",
    "openai>=1.40",
    "tenacity>=8.4",
    "tqdm>=4.66",
    "click>=8.1",
]

[project.optional-dependencies]
sheets = ["google-api-python-client>=2.130", "google-auth>=2.30"]
dev = ["pytest>=8.2", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
sohn-translator = "sohn_translator.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sohn_translator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
